fixed bug regarding wuhan mission reward for japan
Someone complained that he did not get the FO reward for capturing wuhan, because he did so BEFORE capturing Nankin. Since capturing Nankin is the condition for the conquer wuhan mission to start, if wuhan is captured beforehand, the reward for capturing Wuhan will never be granted.
With this change, if the player captures wuhan before Nankin, the reward for Wuhan will be granted upon capturing Nankin.