Hungarian RMI MIO now includes Transport Planes
As suggested by isdvayk on Discord, RMI MIO does not include transport plane type and so if an HUN player wants to build some for any reason, the MIO will not grow as a result. While no such equipment was designed back then by the RMI, if Hungary had wanted to build a transport plane, the RMI would probably have done it, hence my initiative to make this change.